Wireframe Rotation
I created a scene where theres is a wireframe of a cube. When holding
the left grip trigger the cube will match the roation of your hand.
Pressing A resets the rotation. Pressing X and Y will switch between a
wireframe of a cube and a wireframe of a hypercube projection.